Build automation: Parent POM, local build profile (for jar signing), jersey1/2 profiles.
Added a parent POM for inheritance (all common plugins are defined there) and aggregation (automatically build child modules).
GPG jar signing is now only performed when local.build is activated (using local settings.xml)
Jersey 1x/2x dependencies and artifact names are now automatically set based on profiles.

This is only the first part of the automation. We still need to manually set the variable to determine which Jersey profile is used.
Next steps are creating a shell (ant?) script for running both profiles, automatic deployment of artifacts to maven, and perhaps to
BinTray as well (a zip including all the required the dependencies?).